requirements.txt
Flask==3.0.0
pymongo==4.6.1
Werkzeug==3.0.1
itsdangerous==2.1.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
huggingface-hub==0.21.4